pyproject.toml
# SPDX-License-Identifier: Apache-2.0

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-cessda-datasets"
version = "0.1.0"
description = "MCP server for searching CESSDA research datasets"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "CESSDA", email = "cessda@cessda.eu"}
]
keywords = ["mcp", "cessda", "datasets", "research", "data-catalogue"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastmcp>=0.1.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
mcp-cessda-datasets = "mcp_cessda_datasets.server:main"
mcp-cessda-datasets-http = "mcp_cessda_datasets.server_http:main"

[project.urls]
Homepage = "https://github.com/cessda/mcp-cessda-datasets"
Documentation = "https://docs.tech.cessda.eu"
Repository = "https://github.com/cessda/mcp-cessda-datasets"
Issues = "https://github.com/cessda/mcp-cessda-datasets/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
